+++ client/allocrunner/identity_hook_test.go
@@ -5,21 +5,16 @@ package allocrunner
 
 import (
 	"context"
-	"crypto/ed25519"
-	"fmt"
-	"slices"
 	"testing"
 	"time"
 
-	"github.com/go-jose/go-jose/v3"
-	"github.com/go-jose/go-jose/v3/jwt"
 	"github.com/shoenig/test/must"
+
 	"github.com/hashicorp/nomad/ci"
 	"github.com/hashicorp/nomad/client/allocrunner/interfaces"
 	cstructs "github.com/hashicorp/nomad/client/structs"
 	"github.com/hashicorp/nomad/client/widmgr"
 	"github.com/hashicorp/nomad/helper/testlog"
-	"github.com/hashicorp/nomad/helper/uuid"
 	"github.com/hashicorp/nomad/nomad/mock"
 	"github.com/hashicorp/nomad/nomad/structs"
 )
@@ -28,82 +23,6 @@ import (
 var _ interfaces.RunnerPrerunHook = (*identityHook)(nil)
 var _ interfaces.ShutdownHook = (*identityHook)(nil)
 var _ interfaces.TaskStopHook = (*identityHook)(nil)
-
-// MockWIDMgr allows TaskRunner unit tests to avoid having to setup a Server,
-// Client, and Allocation.
-type MockWIDMgr struct {
-	// wids maps identity names to workload identities. If wids is non-nil then
-	// SignIdentities will use it to find expirations or reject invalid identity
-	// names
-	wids  map[string]*structs.WorkloadIdentity
-	key   ed25519.PrivateKey
-	keyID string
-}
-
-func NewMockWIDMgr(wids []*structs.WorkloadIdentity) *MockWIDMgr {
-	_, privKey, err := ed25519.GenerateKey(nil)
-	if err != nil {
-		panic(err)
-	}
-	m := &MockWIDMgr{
-		key:   privKey,
-		keyID: uuid.Generate(),
-	}
-	if wids != nil {
-		m.setWIDs(wids)
-	}
-	return m
-}
-
-// setWIDs is a test helper to use Task.Identities in the MockWIDMgr for
-// sharing TTLs and validating names.
-func (m *MockWIDMgr) setWIDs(wids []*structs.WorkloadIdentity) {
-	m.wids = make(map[string]*structs.WorkloadIdentity, len(wids))
-	for _, wid := range wids {
-		m.wids[wid.Name] = wid
-	}
-}
-func (m *MockWIDMgr) SignIdentities(minIndex uint64, req []*structs.WorkloadIdentityRequest) ([]*structs.SignedWorkloadIdentity, error) {
-	swids := make([]*structs.SignedWorkloadIdentity, 0, len(req))
-	for _, idReq := range req {
-		// Set test values for default claims
-		claims := &structs.IdentityClaims{
-			Namespace:    "default",
-			JobID:        "test",
-			AllocationID: idReq.AllocID,
-			TaskName:     idReq.TaskName,
-		}
-		claims.ID = uuid.Generate()
-		// If test has set workload identities. Lookup claims or reject unknown
-		// identity.
-		if m.wids != nil {
-			wid, ok := m.wids[idReq.IdentityName]
-			if !ok {
-				return nil, fmt.Errorf("unknown identity: %q", idReq.IdentityName)
-			}
-			claims.Audience = slices.Clone(wid.Audience)
-			if wid.TTL > 0 {
-				claims.Expiry = jwt.NewNumericDate(time.Now().Add(wid.TTL))
-			}
-		}
-		opts := (&jose.SignerOptions{}).WithHeader("kid", m.keyID).WithType("JWT")
-		sig, err := jose.NewSigner(jose.SigningKey{Algorithm: jose.EdDSA, Key: m.key}, opts)
-		if err != nil {
-			return nil, fmt.Errorf("error creating signer: %w", err)
-		}
-		token, err := jwt.Signed(sig).Claims(claims).CompactSerialize()
-		if err != nil {
-			return nil, fmt.Errorf("error signing: %w", err)
-		}
-		swid := &structs.SignedWorkloadIdentity{
-			WorkloadIdentityRequest: *idReq,
-			JWT:                     token,
-			Expiration:              claims.Expiry.Time(),
-		}
-		swids = append(swids, swid)
-	}
-	return swids, nil
-}
 
 func TestIdentityHook_Prerun(t *testing.T) {
 	ci.Parallel(t)
@@ -119,7 +38,7 @@ func TestIdentityHook_Prerun(t *testing.T) {
 	}
 
 	alloc := mock.Alloc()
-	task := alloc.Job.TaskGroups[0].Tasks[0]
+	task := alloc.LookupTask("web")
 	task.Identity = wid
 	task.Identities = []*structs.WorkloadIdentity{wid}
 
@@ -129,7 +48,7 @@ func TestIdentityHook_Prerun(t *testing.T) {
 	logger := testlog.HCLogger(t)
 
 	// setup mock signer and WIDMgr
-	mockSigner := NewMockWIDMgr([]*structs.WorkloadIdentity{task.Identity})
+	mockSigner := widmgr.NewMockWIDMgr(task.Identities)
 	mockWIDMgr := widmgr.NewWIDMgr(mockSigner, alloc, logger)
 	allocrunner.widmgr = mockWIDMgr
 	allocrunner.widsigner = mockSigner
@@ -139,7 +58,7 @@ func TestIdentityHook_Prerun(t *testing.T) {
 		{
 			AllocID:      alloc.ID,
 			TaskName:     task.Name,
-			IdentityName: task.Identity.Name,
+			IdentityName: task.Identities[0].Name,
 		},
 	})
 	must.NoError(t, err)
@@ -149,7 +68,7 @@ func TestIdentityHook_Prerun(t *testing.T) {
 	must.Eq(t, hook.Name(), "identity")
 	must.NoError(t, hook.Prerun())
 
-	sid, err := hook.widmgr.Get(cstructs.TaskIdentity{TaskName: task.Name, IdentityName: task.Identity.Name})
+	sid, err := hook.widmgr.Get(cstructs.TaskIdentity{TaskName: task.Name, IdentityName: task.Identities[0].Name})
 	must.Nil(t, err)
 	must.Eq(t, sid.IdentityName, task.Identity.Name)
 	must.NotEq(t, sid.JWT, "")
